requirements.txt
streamlit
streamlit_folium
folium
Pillow
plotly
pandas